+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 4.3.4
 
 - fix: when select a node with click-select, selected combos should be deselected;
+- fix: contextmenu with click trigger does not show the menu up;
 
 #### 4.3.3